Refactor Explorer.js into many components
Test Plan: existing tests, check that UI still works
